Clean up duplicated code and get_usage
