+++ crates/goose/src/providers/databricks.rs
@@ -13,13 +13,13 @@ use tokio_util::io::StreamReader;
 use super::base::{ConfigKey, MessageStream, Provider, ProviderMetadata, ProviderUsage, Usage};
 use super::embedding::EmbeddingCapable;
 use super::errors::ProviderError;
-use super::formats::databricks::{create_request, get_usage, response_to_message};
+use super::formats::databricks::{create_request, response_to_message};
 use super::oauth;
 use super::utils::{get_model, ImageFormat};
 use crate::config::ConfigError;
 use crate::message::Message;
 use crate::model::ModelConfig;
-use crate::providers::formats::openai::response_to_streaming_message;
+use crate::providers::formats::openai::{get_usage, response_to_streaming_message};
 use mcp_core::tool::Tool;
 use serde_json::json;
 use tokio::time::sleep;
@@ -455,13 +455,10 @@ impl Provider for DatabricksProvider {
 
         // Parse response
         let message = response_to_message(response.clone())?;
-        let usage = match response.get("usage").map(get_usage) {
-            Some(usage) => usage,
-            None => {
-                tracing::debug!("Failed to get usage data");
-                Usage::default()
-            }
-        };
+        let usage = response.get("usage").map(get_usage).unwrap_or_else(|| {
+            tracing::debug!("Failed to get usage data");
+            Usage::default()
+        });
         let model = get_model(&response);
         super::utils::emit_debug_trace(&self.model, &payload, &response, &usage);
